Revert "refactor: Integrated new implementations of Icon components"
This reverts commit 5c0e7c6c85dde30cce61c200e828fe30bebde938.
