+++ src/app/components/Logo.tsx
@@ -1,7 +1,7 @@
 import React from "react";
+import LogoIcon from "./icon/LogoIcon";
 import { cva, VariantProps } from "class-variance-authority";
 import { cn } from "@/lib/utils";
-import Icon from "./icon/Icon";
 
 interface LogoProps extends VariantProps<typeof logoVariants> {}
 
@@ -20,7 +20,7 @@ export const logoVariants = cva("flex items-center max-w-fit font-medium text-ba
 export const Logo: React.FC<LogoProps> = ({ variant, textVariant: text }) => {
   return (
     <div className={cn(logoVariants({ variant }))}>
-      <Icon.Logo className="w-[46px] h-[46px]"></Icon.Logo>
+      <LogoIcon className="w-[46px] h-[46px]"></LogoIcon>
       <div className={cn(logoVariants({ textVariant: text }))}>P2PCraft</div>
     </div>
   );
