+++ src/app/components/navigation/Footer.tsx
@@ -1,24 +1,25 @@
 import React from "react";
-import Button from "../button/Button";
+import LogoIcon from "../icon/LogoIcon";
+import Button from "../Button";
 import { cva, VariantProps } from "class-variance-authority";
 import { cn } from "@/lib/utils";
-import Icon from "../icon/Icon";
+import GithubIcon from "../icon/GithubIcon";
+import YoutubeIcon from "../icon/YoutubeIcon";
+import DiscordIcon from "../icon/DiscordIcon";
 
-interface FooterProps extends VariantProps<typeof footerVariants> {}
+interface FooterProps extends VariantProps<typeof footerVariants> { }
 
-const footerVariants = cva(
-  "container flex mx-auto mt-auto items-center justify-between px-4 w-full text-black-900",
-  {
-    variants: {
-      variant: {
-        primary: "",
-        dark: "bg-dark-green-900",
-      },
+const footerVariants = cva("container flex mx-auto mt-auto items-center justify-between px-4 w-full text-black-900", {
+  variants: {
+    variant: {
+      primary: "",
+      dark: "bg-dark-green-900",
     },
-  }
-);
+  },
+});
 
 export const Footer: React.FC<FooterProps> = ({ variant = "primary" }: FooterProps) => {
+
   return (
     <div className={cn(footerVariants({ variant }))}>
       <FooterLogo></FooterLogo>
@@ -28,9 +29,9 @@ export const Footer: React.FC<FooterProps> = ({ variant = "primary" }: FooterPro
         <Button variant="link">FAQ</Button>
       </div>
       <div className="flex">
-        <Icon.Discord className="w-[78px] h-[78px]" />
-        <Icon.Youtube className="w-[78px] h-[78px]" />
-        <Icon.Github className="w-[78px] h-[78px]" />
+        <DiscordIcon width={78} height={78} />
+        <YoutubeIcon width={78} height={78} />
+        <GithubIcon width={78} height={78} />
       </div>
     </div>
   );
@@ -40,7 +41,7 @@ export const FooterLogo = () => {
   return (
     <>
       <div className="flex items-center">
-        <Icon.Logo className="w-16 h-16" />
+        <LogoIcon className="w-16 h-16" />
         <div className="md:min-w-fit text-black-900">© 2024 P2PCraft</div>
       </div>
     </>
