+++ src/app/components/navigation/NavBar.tsx
@@ -1,22 +1,19 @@
 import React from "react";
 import { Logo, logoVariants } from "../Logo";
-import Button, { buttonVariants } from "../button/Button";
+import Button, { buttonVariants } from "../Button";
 import { cva, VariantProps } from "class-variance-authority";
 import { cn } from "@/lib/utils";
 
 interface NavbarProps extends VariantProps<typeof navbarVariants> {}
 
-const navbarVariants = cva(
-  "container mx-auto px-4 top-0 py-1 sticky flex justify-between items-center",
-  {
-    variants: {
-      variant: {
-        primary: "",
-        dark: "bg-dark-green-900",
-      },
+const navbarVariants = cva("container mx-auto px-4 top-0 py-1 sticky flex justify-between items-center", {
+  variants: {
+    variant: {
+      primary: "",
+      dark: "bg-dark-green-900",
     },
-  }
-);
+  },
+});
 
 export const NavBar: React.FC<NavbarProps> = ({ variant = "primary" }: NavbarProps) => {
   let buttonVariant: VariantProps<typeof buttonVariants>["textColor"];
